+++ aioamqp/protocol.py
@@ -56,7 +56,7 @@ class AmqpProtocol(asyncio.StreamReaderProtocol):
         logger.warning("Connection lost exc=%r", exc)
         self.connection_closed.set()
         self.is_open = False
-        self._close_channels(reply_code=None, reply_text='Connection lost')
+        self._close_channels(exception=exc)
         super().connection_lost(exc)
 
     @asyncio.coroutine
@@ -192,18 +192,37 @@ class AmqpProtocol(asyncio.StreamReaderProtocol):
         """
         self.channels_ids.append(channel_id)
 
-    def _close_channels(self, reply_code, reply_text):
+    def _close_channels(self, reply_code=None, reply_text=None, exception=None):
+        """Cleanly close channels
+
+            Args:
+                reply_code:     int, the amqp error code
+                reply_text:     str, the text associated to the error_code
+                exc:            the exception responsible of this error
+
+        """
+        if exception is None:
+            exception = exceptions.ChannelClosed(reply_code, reply_text)
+
+        if self._on_error_callback:
+            if asyncio.iscoroutinefunction(self._on_error_callback):
+                asyncio.async(self._on_error_callback(exception))
+            else:
+                self._on_error_callback(exceptions.ChannelClosed(exception))
+
         for channel in self.channels.values():
-            channel.connection_closed(reply_code, reply_text)
+            channel.connection_closed(reply_code, reply_text, exception)
 
     @asyncio.coroutine
     def run(self):
         while not self.stop_now.done():
             try:
                 yield from self.dispatch_frame()
-            except exceptions.AmqpClosedConnection:
+            except exceptions.AmqpClosedConnection as exc:
                 logger.info("Close connection")
                 self.stop_now.set_result(None)
+
+                self._close_channels(exception=exc)
             except Exception:
                 logger.exception('error on dispatch')
 
@@ -251,11 +270,6 @@ class AmqpProtocol(asyncio.StreamReaderProtocol):
             reply_text, reply_code, class_id, method_id)
         self._close_channels(reply_code, reply_text)
 
-        if self._on_error_callback:
-            if asyncio.iscoroutinefunction(self._on_error_callback):
-                yield from self._on_error_callback(exceptions.ChannelClosed(reply_code, reply_text))
-            else:
-                self._on_error_callback(exceptions.ChannelClosed(reply_code, reply_text))
 
     @asyncio.coroutine
     def tune(self, frame):