+++ aioamqp/tests/test_connection_lost.py
@@ -1,17 +1,33 @@
 import unittest
+import unittest.mock
 import asyncio
 
 from . import testcase
 from . import testing
+
+from .. import exceptions
 
 
 class ConnectionLostTestCase(testcase.RabbitTestCase, unittest.TestCase):
 
     _multiprocess_can_split_ = True
 
+    def setUp(self):
+        super().setUp()
+        self.callback_called = False
+
+    def tearDown(self):
+        super().setUp()
+        self.callback_called = False
+
     @testing.coroutine
     def test_connection_lost(self):
+
+        def callback(*args, **kwargs):
+            self.callback_called = True
+
         amqp = self.amqp
+        amqp._on_error_callback = callback
         transport = self.transport
         channel = self.channel
         self.assertTrue(amqp.is_open)
@@ -21,3 +37,4 @@ class ConnectionLostTestCase(testcase.RabbitTestCase, unittest.TestCase):
         self.assertFalse(amqp.is_open)
         self.assertFalse(channel.is_open)
         yield from amqp.close()
+        self.assertTrue(self.callback_called)